--- src/components/errors/error.500.njk
@@ -1,11 +1,0 @@
-{% extends "../../layouts/govuk.njk" %}
-
-{% block pageTitle %}
-  Internal Server Error
-{% endblock %}
-
-{% block content %}
-  {{ super() }}
-  <h1 class="govuk-heading-xl">Sorry an error occurred</h1>
-  <p class="govuk-body">{{ errorMessage | default('Something went wrong while processing the request.') }}</p>
-{% endblock %}
